+++ toshl/client.py
@@ -91,6 +91,12 @@ class Account(object):
         response = self.client._make_request('/accounts/{0}'.format(account_id))
         return response.json()
 
+    def create(self, json_payload):
+        response = self.client._make_request(
+            '/accounts', 'POST', json=json_payload)
+        if response.status_code == 201:
+            return self.client._parse_location_header(response)
+
 
 class Category(object):
     def __init__(self, client):