+++ CHANGELOG.md
@@ -5,5 +5,6 @@
 ### v0.0.4
 
 - Add get() method to Account
+- Add create() method to Account
 - Add proper testing for ToshlException
 - Improve string representation of ToshlException